refactor: 将物品选择器抽离
