+++ src/views/WorkflowPage.vue
@@ -1,8 +1,7 @@
 <script setup lang="ts">
-import { computed, inject, onMounted, onBeforeUnmount, ref, watch, h, type Ref } from 'vue'
+import { computed, inject, onMounted, onBeforeUnmount, ref, watch, type Ref } from 'vue'
 import {
-  NBackTop, NButton, NFloatButton, NFloatButtonGroup, NIcon, NInputGroup, NInputGroupLabel, NSelect, NTabs, NTabPane, NTooltip, 
-  type SelectOption, type SelectRenderLabel,
+  NBackTop, NButton, NFloatButton, NFloatButtonGroup, NIcon, NInputGroup, NInputGroupLabel, NTabs, NTabPane, NTooltip, 
   useMessage
 } from 'naive-ui'
 import {
@@ -15,7 +14,7 @@ import {
   UnfoldMoreSharp, UnfoldLessSharp,
 } from '@vicons/material'
 import FoldableCard from '@/components/templates/FoldableCard.vue'
-import ItemSpan from '@/components/custom/item/ItemSpan.vue'
+import ItemSelector from '@/components/custom/item/ItemSelector.vue'
 import ItemSelectTable from '@/components/custom/item/ItemSelectTable.vue'
 import CraftStatistics from '@/components/custom/general/CraftStatistics.vue'
 import CraftStatements from '@/components/custom/general/CraftStatements.vue'
@@ -23,7 +22,6 @@ import CraftStatementsPro from '@/components/custom/general/CraftStatementsPro.v
 import ModalCostAndBenefit from '@/components/modals/ModalCostAndBenefit.vue'
 import ModalPreferences from '@/components/modals/ModalPreferences.vue'
 import { useStore } from '@/store'
-import { XivUnpackedItems } from '@/assets/data'
 import {
   getDefaultWorkflow, fixWorkState, _VAR_MAX_WORKFLOW
 } from '@/models/workflow'
@@ -177,49 +175,6 @@ const pageHeightVals = computed(() => {
 })
 
 // #region content-items
-const itemOptions = computed(() => {
-  return Object.values(XivUnpackedItems).filter(item => item.rids?.length > 0).map(item => {
-    return {
-      label: item.lang[0],
-      value: item.id
-    }
-  })
-})
-const renderItemLabel : SelectRenderLabel = (option) => {
-  if (!option.value || typeof option.value !== 'number') {
-    return h('div', null, [option.label as string])
-  }
-  return h(ItemSpan, {
-    itemInfo: getItemInfo(option.value),
-  })
-}
-const filterItem = (pattern: string, option: SelectOption) => {
-  if (!pattern) {
-    return true
-  }
-  if (!option.value || typeof option.value !== 'number') {
-    return false
-  }
-  const item = getItemInfo(option.value)
-
-  let itemMatched = false
-  const availableKeywords = [
-    item.name_zh, item.name_en, item.name_ja
-  ]
-  availableKeywords.forEach(keyword => {
-    if (keyword?.toLowerCase().includes(pattern.toLowerCase())) {
-      itemMatched = true
-    }
-  })
-  if (itemMatched) return true
-
-  if (item.id.toString() === pattern) return true
-  if (item.itemLevel.toString() === pattern) return true
-  if (item.patch === pattern) return true
-
-  return false
-}
-const itemInputVal = ref<number | null>(null)
 const handleItemInputValueUpdate = (value: number) => {
   if (!value) return
   if (currentWorkflow.value.targetItems[value]) {
@@ -228,7 +183,6 @@ const handleItemInputValueUpdate = (value: number) => {
     currentWorkflow.value.targetItems[value] = 1
     currentWorkflow.value.preparedItems.craftTarget[value] = 0
   }
-  itemInputVal.value = null
 }
 const handleClearCurrentWorkflow = () => {
   currentWorkflow.value.targetItems = {}
@@ -475,15 +429,8 @@ const handleAnalysisItemPrices = async () => {
           <div class="top-actions">
             <n-input-group>
               <n-input-group-label>{{ t('添加物品') }}</n-input-group-label>
-              <n-select
-                v-model:value="itemInputVal"
-                filterable
-                :filter="filterItem"
-                :options="itemOptions"
-                :render-label="renderItemLabel"
-                :placeholder="t('支持按物品名/ID/品级/版本搜索')"
-                :title="t('支持按物品名/ID/品级/版本搜索')"
-                @update:value="handleItemInputValueUpdate"
+              <ItemSelector
+                @on-item-selected="handleItemInputValueUpdate"
               />
             </n-input-group>
           </div>
